.gitignore .vscode/
Added settings.json to .vscode/ (really helps in file lookup/search
filters, like for the libs folder), looking for a way to backup the
configs outside of this git repo because they are integral to my
productivity and I don’t want to have to research everything again.

Also added new line to end of file.